Move core into "core" subdirectory. That way we can clearly see what goes where in the future.
