+++ core/src/typescript/main/core/api/LogChannel.ts
