+++ core/src/typescript/main/core/api/LogLevel.ts
