+++ core/src/typescript/main/core/api/LogMessage.ts
