+++ core/src/typescript/main/core/api/LogProvider.ts
