+++ core/src/typescript/main/core/api/LogRuntime.ts
