+++ core/src/typescript/main/core/api/RawLogChannel.ts
