+++ core/src/typescript/main/core/api/RawLogMessage.ts
