+++ core/src/typescript/main/core/api/config/LogSettings.ts
