+++ core/src/typescript/main/core/api/type/ExceptionType.ts
