+++ core/src/typescript/main/core/api/type/LogChannelType.ts
