+++ core/src/typescript/main/core/api/type/LogMessageType.ts
