+++ core/src/typescript/main/core/api/type/LoggerNameType.ts
