+++ core/src/typescript/main/core/impl/LogRuntimeImpl.ts
