+++ core/src/typescript/main/core/impl/LoggerImpl.ts
